Moved ignored rules to PolicyContext for easier organization
